added GINT_TO_POINTER and friends, fixed format strings, for 64-bitness.
2003-01-14  Manish Singh  <yosh@gimp.org>

        * many files in app, libgimp, and plug-ins: added GINT_TO_POINTER and
        friends, fixed format strings, for 64-bitness.

        * many Makefile.am's in plug-ins: use LDFLAGS = @LDFLAGS@ $(mwindows)
        so any user set flags get passed on

        * app/plug_in.h: removed unused structure members housed in #ifdef
        G_OS_WIN32.

+++ app/docindex.c
@@ -261,7 +261,7 @@ save_to_ideas (gpointer data,
 
   title = GTK_LABEL (GTK_BIN (data)->child)->label;
 
-  fprintf ((FILE *) user_data, "%d %s\n", strlen (title), title);
+  fprintf ((FILE *) user_data, "%d %s\n", (int) strlen (title), title);
 }
 
 static void
@@ -272,7 +272,7 @@ save_list_to_ideas (gpointer data,
 
   title = (gchar *) data;
 
-  fprintf ((FILE *) user_data, "%d %s\n", strlen (title), title);
+  fprintf ((FILE *) user_data, "%d %s\n", (int) strlen (title), title);
 }
 
 static void
