+++ plug-ins/script-fu/siod/regex.c
@@ -134,7 +134,7 @@ regex_prin1 (LISP ptr, struct gen_printio *f)
   regex_t *p;
   p = get_tc_regex (ptr)->r;
   sprintf (buffer, "#<REGEX %p nsub=%d",
-	   p, p->re_nsub);
+	   p, (int) p->re_nsub);
   gput_st (f, buffer);
   gput_st (f, ">");
 }
